+++ src/calculated-property-rule.ts
@@ -1,4 +1,4 @@
-import { Rule, RuleOptions } from "./rule";
+import { Rule, RuleInvocationOptions, RuleOptions } from "./rule";
 import { Type } from "./type";
 import { Property, Property$init, Property$pendingInit, PropertyRuleOptions } from "./property";
 import { Entity } from "./entity";
@@ -14,8 +14,7 @@ export class CalculatedPropertyRule extends Rule {
 
 	// Public settable properties that are simple values with no side-effects or logic
 	defaultIfError: any;
-
-	isDefaultValue: boolean;
+	private isDefaultValue: boolean;
 
 	// Backing fields for properties that are settable and also derived from
 	// other data, calculated in some way, or cannot simply be changed
@@ -33,6 +32,9 @@ export class CalculatedPropertyRule extends Rule {
 		if (options) {
 			if (options.property) {
 				property = typeof options.property === "string" ? rootType.getProperty(options.property) as Property : options.property as Property;
+
+				if (options.isDefaultValue && property.isList)
+					(options as RuleInvocationOptions).onInitNew = true;
 
 				// indicate that the rule is responsible for returning the value of the calculated property
 				options.returns = [property];
@@ -58,7 +60,7 @@ export class CalculatedPropertyRule extends Rule {
 		// Public settable properties
 		this.defaultIfError = defaultIfError;
 
-		this.isDefaultValue = options.isDefaultValue;
+		this.isDefaultValue = !!options.isDefaultValue;
 
 		// Backing fields for properties
 		if (calculateFn) Object.defineProperty(this, "_calculateFn", { enumerable: false, value: calculateFn, writable: true });
@@ -143,7 +145,8 @@ export class CalculatedPropertyRule extends Rule {
 			const newList = newValue;
 
 			// ensure the initial calculation of the list does not raise change events
-			if (Property$pendingInit(obj, this.property))
+			// defaulting a list property should raise change events
+			if (!this.isDefaultValue && Property$pendingInit(obj, this.property))
 				Property$init(this.property, obj, newList);
 			else {
 				// compare the new list to the old one to see if changes were made